adjust on_disabled type

+++ pallets/collator-selection/src/mock.rs
@@ -161,7 +161,7 @@ impl pallet_session::SessionHandler<u64> for TestSessionHandler {
 		SessionHandlerCollators::set(keys.into_iter().map(|(a, _)| *a).collect::<Vec<_>>())
 	}
 	fn on_before_session_ending() {}
-	fn on_disabled(_: u32) {}
+	fn on_disabled(_: usize) {}
 }
 
 parameter_types! {
